+++ HISTORY.md
@@ -10,6 +10,7 @@
     * TTS library: new "started speech response?" predicate
     * "scn" extension category for cloned scenes and tutorials
     * new "scn_scale([num])" extension, scales and animates a launched tutorial, reports the scale if given no number or zero
+    * new "scn_exit" extension, closes the tutorial dialog and redisplays the scenes in the corral
 * **Notable Changes:**
     * always "normalize" SVGs on import, avoids "cut-off" costume parts
 * **Translation Updates:**
@@ -18,6 +19,7 @@
 ### 2025-12-04
 * extensions: new "scn" category for cloned scenes and tutorials
 * extensions: new "scn_scale([num])" extension, scales and animates a launched tutorial, reports the scale if given no number or zero
+* gui, extensions: new "scn_exit" extension, closes the tutorial dialog and redisplays the scenes in the corral
 
 ### 2025-12-03
 * gui: always normalize SVGs on import